// src/interfaces/IPrescriptionsResponse.response.interface.ts
// Generated by https://quicktype.io

import { IPrescriptionToSupply } from "./IPrescription.interface";

export interface IPrescriptionResponse {
  data: Data;
  ok: boolean;
}

export interface IPrescriptionWithMedicinesResponse {
  data: PrescriptionData;
  ok: boolean;
}

export interface PrescriptionData {
  prescription: IPrescriptionToSupply;
}

export interface Data {
  prescriptions: Prescription[];
}

export interface Prescription {
  id: string;
  user_id: UserID;
  user: User;
  prescription_status_id: string;
  prescription_status: PrescriptionStatus;
  folio: number;
  patient_name: string;
  observations: string;
  instructions: Instructions;
  created_at: string;
  updated_at: string;
  supplied_at: null;
  deleted_at: null;
}

export enum Instructions {
  Asdf = "asdf",
  Asdfasdf = "asdfasdf",
  IbuprofenoTomarCada8Horas = "Ibuprofeno: tomar cada 8 horas",
  InstructionsASDFASDF = "ASDFASDF",
}

export interface PrescriptionStatus {
  id: string;
  name: PrescriptionStatusName;
  deleted_at: null;
}

export enum PrescriptionStatusName {
  Pendiente = "Pendiente",
}

export interface User {
  id: UserID;
  role_id: string;
  role: Role;
  name: UserName;
  surname: Surname;
  full_name: FullName;
  deleted_at: string;
}

export enum FullName {
  CECANADMINSuper = "CECAN ADMIN super",
}

export enum UserID {
  Can100481 = "CAN100481",
}

export enum UserName {
  CecanAdmin = "CECAN ADMIN",
}

export interface Role {
  id: string;
  name: RoleName;
}

export enum RoleName {
  Admin = "Admin",
}

export enum Surname {
  Super = "super",
}